Removed old check for if X button was clicked
The old handling for the show image windows has been removed in favor
of imshow coupled with waitKey(0). This was due to issues with the image
not appearing when running on the Raspberry Pi 0W. As the image should
only be shown when running the OCR module seperatly, no issues should
be encountered when running the exporter.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>